refactor(memory): move documented memory defaults to constants.py
